chore: change aas-lib groupId
