+++ aas-lib/src/main/java/de/fraunhofer/iosb/ilt/aas/lib/auth/impl/NoAuth.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.aas.lib.auth.impl;
+package de.fraunhofer.iosb.ilt.aas.lib.auth.impl;
 
 
-import de.fraunhofer.iosb.aas.lib.auth.AuthenticationMethod;
+import de.fraunhofer.iosb.ilt.aas.lib.auth.AuthenticationMethod;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
