+++ aas-lib/src/main/java/de/fraunhofer/iosb/ilt/aas/lib/net/AasAccessUrl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.aas.lib.net;
+package de.fraunhofer.iosb.ilt.aas.lib.net;
 
 import org.jetbrains.annotations.NotNull;
